move allof
