+++ packages/hono-takibi/src/generators/zod/generate-zod-object.ts
@@ -1,7 +1,7 @@
 import type { Schema } from '../../types'
 import type { Config } from '../../config'
 import { generateZodRecordSchema } from './generate-zod-record-schema'
-import { generateAllOfCode } from '../openapi/components/allof/generate-allof-code'
+import { generateAllOfCode } from '../hono/openapi/component/allof/generate-allof-code'
 import { generateOneOfCode } from '../openapi/components/oneof/generate-oneof-code'
 import { generateAnyOfCode } from '../openapi/components/anyof/generate-anyof-code'
 import { generateZodPropertiesSchema } from './generate-zod-properties-schema'
