+++ packages/hono-takibi/src/generators/zod/schema/generate-zod-schema.ts
@@ -5,7 +5,7 @@ import { generateZodStringSchema } from '../generate-zod-string-schema'
 import { isFormatString } from '../../../core/validator/is-format-string'
 import { generateZodNumberSchema } from '../generate-zod-number-schema'
 import { generateZodIntegerSchema } from '../generate-zod-integer-schema'
-import { generateAllOfCode } from '../../openapi/components/allof/generate-allof-code'
+import { generateAllOfCode } from '../../hono/openapi/component/allof/generate-allof-code'
 import { generateAnyOfCode } from '../../openapi/components/anyof/generate-anyof-code'
 import { generateOneOfCode } from '../../openapi/components/oneof/generate-oneof-code'
 import { getVariableSchemaNameHelper } from '../../../core/helper/get-variable-schema-name-helper'
